[BUG] AWS CodeDeploy AllowTraffic 이벤트에서 무한 로딩 문제
